pyproject.toml
[project]
name = "translate"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = [
    "openai-whisper",
    "yt-dlp",
]
